// src/model/enums/SDKTokenType.ts
//
//  SDKTokenType.ts
//  Supernova SDK
//
//  Created by Jiri Trecak.
//  Copyright © 2021 Supernova. All rights reserved.
//

// --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- --- ---
// MARK: - Token type enum

export enum TokenType {
  color = 'Color',
  typography = 'Typography',
  radius = 'Radius',
  font = 'Font',
  measure = 'Measure',
  shadow = 'Shadow',
  border = 'Border',
  gradient = 'Gradient',
  text = 'Text',
  blur = 'Blur',
  generic = 'GenericToken'
}